2 var

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -66,6 +66,11 @@ public class AdminController {
 
     @DeleteMapping("/admin/users/{id}/delete")
     public String deleteUser(@PathVariable("id") Long id) {
+        User user = userService.findUserById(id);
+        if (user == null) {
+            throw new RuntimeException("There is no user with ID = " +
+                    id + " in Database");
+        }
         userService.deleteUser(id);
         return "redirect:/admin/users";
     }
